+++ backend/garpix_notify/admin/config.py
@@ -20,8 +20,10 @@ class NotifyConfigAdmin(SingletonModelAdmin):
             'fields': ('call_url_type', 'call_api_id', 'call_login', 'call_password')
         }),
         ('Telegram', {
-            'fields': ('telegram_api_key', 'telegram_bot_name', 'telegram_welcome_text', 'telegram_help_text', 'telegram_bad_command_text',
-                       'telegram_success_added_text', 'telegram_failed_added_text')
+            'fields': ('telegram_api_key', 'telegram_bot_name', 'telegram_welcome_text', 'telegram_help_text',
+                       'telegram_bad_command_text', 'telegram_success_added_text', 'telegram_failed_added_text',
+                       'telegram_parse_mode', 'telegram_disable_notification', 'telegram_disable_web_page_preview',
+                       'telegram_allow_sending_without_reply', 'telegram_timeout')
         }),
         ('Viber', {
             'fields': ('viber_api_key', 'viber_bot_name', 'viber_success_added_text',