+++ backend/garpix_notify/models/config.py
@@ -1,20 +1,12 @@
 from django.db import models
 from solo.models import SingletonModel
 
+from garpix_notify.models.choices import EMAIL_MALLING, PARSE_MODE_TELEGRAM
+
 
 class NotifyConfig(SingletonModel):
-    class EMAIL_MALLING:
-        CC = 0
-        BCC = 1
-        TYPES = (
-            (CC, 'Обычная рассылка'),
-            (BCC, 'Скрытая рассылка'),
-        )
-
     class SMS_URL:
-        """
-        URL СМС провайдера
-        """
+        """ URL СМС провайдера """
 
         SMSRU_ID = 0
         WEBSZK_ID = 1
@@ -43,9 +35,7 @@ class NotifyConfig(SingletonModel):
         )
 
     class CALL_URL:
-        """
-        URL Оператора связи
-        """
+        """ URL Оператора связи """
         SMSRU_CALL_API_ID = 0
         SMSRU_CALL_ID = 1
         SMSCENTRE_ID = 2
@@ -105,6 +95,17 @@ class NotifyConfig(SingletonModel):
     telegram_failed_added_text = models.TextField(blank=True,
                                                   default='Ошибка при привязке учетной записи. Пожалуйста, свяжитесь с техподдержкой',
                                                   verbose_name='Telegram - Текст провал, не добавлен код')
+    telegram_parse_mode = models.CharField(default=PARSE_MODE_TELEGRAM.EMPTY, choices=PARSE_MODE_TELEGRAM.TYPES,
+                                           verbose_name='Тип парсера телеграм сообщений', max_length=100, blank=True)
+    telegram_disable_notification = models.BooleanField(verbose_name='Пользователи получат уведомление без звука',
+                                                        default=False)
+    telegram_disable_web_page_preview = models.BooleanField(
+        verbose_name='Отключает предварительный просмотр ссылок в сообщениях',
+        default=False)
+    telegram_allow_sending_without_reply = models.BooleanField(
+        verbose_name='Разрешить, если сообщение должно быть отправлено, даже если ответное сообщение не найдено',
+        default=False)
+    telegram_timeout = models.FloatField(default=None, blank=True, verbose_name="Тайм-аут чтения с сервера", null=True)
     viber_api_key = models.CharField(default='000000000:AAAAAAAAAA-AAAAAAAA-_AAAAAAAAAAAAAA', blank=True,
                                      max_length=255, verbose_name='Viber API Key')
     viber_bot_name = models.CharField(blank=True, max_length=255, verbose_name='Название viber бота',