+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added the Call method (Notify.call(phone, user, url, **kwargs)),
 - Fix bugs,
 - Fixed work NOTIFY_USER_WANT_MESSAGE_CHECK,
+- New settings for Telegram,
 - Telecom operators added.
 
 ### 5.7.0 (01.06.2022)